+++ Changelog.md
@@ -2,7 +2,8 @@
 
 ## 4.0.0 (next)
 
-* `ResultSetMetadata::col_display_size` now returns `Option<NonZeroUsize>`, instead of `isize`.
+* `ResultSetMetadata::col_display_size` now returns `Option<NonZeroUsize>` instead of `isize`.
+* `ResultSetMetadata::col_octet_length` now returns `Option<NonZeroUsize>`, instead of `isize`.
 * `ResultSetMetadata::utf8_display_sizes` now iterates over `Option<NonZeroUsize>`, instead of `usize`.
 * `DataType` now stores the length of variadic types as `Option<NonZeroUsize`> insteaf of `usize`.
 * `DataType::display_size` now returns `Option<NonZeroUSize>` instead of `Option<usize>`.